requirements-dev.txt
black==25.1.0
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
filelock==3.18.0
identify==2.6.9
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pre_commit==4.2.0
pyserial==3.5
PyYAML==6.0.2
setuptools==77.0.3
virtualenv==20.29.3
wheel==0.45.1
